split out keychain tests from SecureItemStorage tests
